[TASK] Remove dev dependency `"ergebnis/composer-normalize"`
The hyped `"ergebnis/composer-normalize"` composer plugin to
normalize composer.json file and using it as a pipeline check
operats strictly on semver and not allowing a more practical
approach regardin the version handling of dependencies.

Due to the fact that having another mindset then this plugin,
it's not usable until it can be controlled. [1]

Therefore, the plugin is removed again until this point which
may never come. Handling the composer.json file manually in
shape is doable.

Used command(s):

```terminal
> Build/Scripts/runTests.sh -s composer \
    -- remove --dev "ergebnis/composer-normalize"
```

[1] https://github.com/ergebnis/composer-normalize/issues/1287

Releases: main