[WIP][FEATURE] Add configurable shortcut and mointpoint redirect statuscode
todo's:

* add `redirect_statuscode` field next to shortcut or mointpoint
  configuration.
* add tests.

----------------------------------------------------------------

TYPO3 uses conservative hardcoded `HTTP Status Codes` for the
`Shortcut` redirects (307) and `Mountpoint` redirects (303),
which does not always meets the requirements for all instances.

This change adds the ability to configure the statuscode for
`shortcut` and `mountpoint` redirects by an editor, defaulting
to take the default value. Therefore, the redirect service is
modified to respect the configured value and using the known
hardcoded default as fallback value.

Resolves: #?
Releases: main